requirements.txt
# Core app dependencies
streamlit==1.33.0
python-dotenv==1.0.1

# LangChain core and modules
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.47
langchain-groq==0.0.2
langchain-google-genai==0.0.8

# Vector store and embedding
faiss-cpu==1.7.4
sentence-transformers==2.5.1
huggingface-hub==0.20.3
transformers==4.37.2

# Text processing
regex==2024.4.16
tiktoken==0.6.0

# Search and scraping tools
duckduckgo-search==4.1
wikipedia==1.4.0

# Document parsing
pypdf==4.1.0
pdfplumber==0.10.3

# Memory and chaining
pydantic==2.6.1
typing-extensions==4.10.0

# Optional (for enhanced caching or async performance)
aiohttp==3.9.3
httpx==0.26.0

